+++ packages/routed/CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## 0.5.1 - 2026-08-25
 
 - Adopted `very_good_analysis` and completed public API documentation for the
   batteries-included facade.